vocabulary types endpoint: split implementation for service, config and results in different files and cleaned up code.
